+++ graphql-server/src/scalars.ts
@@ -7,5 +7,6 @@ export const Int = GraphQLInt
 export const Float = GraphQLFloat
 export const DateTime = customScalars.DateTime
 export const BigInteger = customScalars.BigInt
+export const BigDecimalScalar = customScalars.BigDecimal
 export const Bytes = customScalars.Bytes
 export const Json = customScalars.JSON