+++ typeorm-codegen/src/codegen.ts
@@ -1,9 +1,9 @@
-import type {Entity, Enum, JsonObject, Model, Prop, Union} from "@subsquid/openreader/lib/model"
-import {unexpectedCase} from "@subsquid/util-internal"
-import {OutDir, Output} from "@subsquid/util-internal-code-printer"
-import {toCamelCase} from "@subsquid/util-naming"
-import assert from "assert"
-import * as path from "path"
+import type {Entity, Enum, JsonObject, Model, Prop, Union} from '@subsquid/openreader/lib/model'
+import {unexpectedCase} from '@subsquid/util-internal'
+import {OutDir, Output} from '@subsquid/util-internal-code-printer'
+import {toCamelCase} from '@subsquid/util-naming'
+import assert from 'assert'
+import * as path from 'path'
 
 
 export function generateOrmModels(model: Model, dir: OutDir): void {
@@ -59,17 +59,32 @@ export function generateOrmModels(model: Model, dir: OutDir): void {
                             out.line('@PrimaryColumn_()')
                         } else {
                             addIndexAnnotation(entity, key, imports, out)
-                            if (prop.type.name === 'BigInt') {
-                                imports.useMarshal()
-                                out.line(
-                                    `@Column_("numeric", {transformer: marshal.bigintTransformer, nullable: ${prop.nullable}})`
-                                )
-                            } else {
-                                out.line(
-                                    `@Column_("${getDbType(prop.type.name)}", {nullable: ${
-                                        prop.nullable
-                                    }})`
-                                )
+                            switch (prop.type.name) {
+                                case 'BigInt':
+                                    imports.useMarshal()
+                                    out.line(
+                                        `@Column_("${getDbType(prop.type.name)}", {transformer: marshal.bigintTransformer, nullable: ${prop.nullable}})`
+                                    )
+                                    break
+                                case 'BigDecimal':
+                                    imports.useMarshal()
+                                    out.line(
+                                        `@Column_("${getDbType(prop.type.name)}", {transformer: marshal.bigdecimalTransformer, nullable: ${prop.nullable}})`
+                                    )
+                                    break
+                                case 'Float':
+                                    imports.useMarshal()
+                                    out.line(
+                                        `@Column_("${getDbType(prop.type.name)}", {transformer: marshal.floatTransformer, nullable: ${prop.nullable}})`
+                                    )
+                                    break
+                                default:
+                                    out.line(
+                                        `@Column_("${getDbType(prop.type.name)}", {nullable: ${
+                                            prop.nullable
+                                        }})`
+                                    )
+                                    break
                             }
                         }
                         break
@@ -161,7 +176,7 @@ export function generateOrmModels(model: Model, dir: OutDir): void {
                     default:
                         throw unexpectedCase((prop.type as any).kind)
                 }
-                out.line(`${key}!: ${getPropJsType('entity', prop)}`)
+                out.line(`${key}!: ${getPropJsType(imports, 'entity', prop)}`)
             }
         })
         out.write()
@@ -181,6 +196,7 @@ export function generateOrmModels(model: Model, dir: OutDir): void {
             case 'DateTime':
                 return 'timestamp with time zone'
             case 'BigInt':
+            case 'BigDecimal':
                 return 'numeric'
             case 'Bytes':
                 return 'bytea'
@@ -207,7 +223,7 @@ export function generateOrmModels(model: Model, dir: OutDir): void {
             for (const key in object.properties) {
                 const prop = object.properties[key]
                 importReferencedModel(imports, prop)
-                out.line(`private _${key}!: ${getPropJsType('object', prop)}`)
+                out.line(`private _${key}!: ${getPropJsType(imports, 'object', prop)}`)
             }
             out.line()
             out.block(
@@ -226,14 +242,14 @@ export function generateOrmModels(model: Model, dir: OutDir): void {
                 const prop = object.properties[key]
                 out.line()
                 printComment(prop, out)
-                out.block(`get ${key}(): ${getPropJsType('object', prop)}`, () => {
+                out.block(`get ${key}(): ${getPropJsType(imports, 'object', prop)}`, () => {
                     if (!prop.nullable) {
                         out.line(`assert(this._${key} != null, 'uninitialized access')`)
                     }
                     out.line(`return this._${key}`)
                 })
                 out.line()
-                out.block(`set ${key}(value: ${getPropJsType('object', prop)})`, () => {
+                out.block(`set ${key}(value: ${getPropJsType(imports, 'object', prop)})`, () => {
                     out.line(`this._${key} = value`)
                 })
             }
@@ -382,11 +398,14 @@ export function generateOrmModels(model: Model, dir: OutDir): void {
 }
 
 
-function getPropJsType(owner: 'entity' | 'object', prop: Prop): string {
+function getPropJsType(imports: ImportRegistry, owner: 'entity' | 'object', prop: Prop): string {
     let type: string
     switch(prop.type.kind) {
         case 'scalar':
             type = getScalarJsType(prop.type.name)
+            if (type == 'BigDecimal') {
+                imports.useBigDecimal()
+            }
             break
         case 'enum':
         case 'object':
@@ -407,7 +426,7 @@ function getPropJsType(owner: 'entity' | 'object', prop: Prop): string {
             type = prop.type.entity + '[]'
             break
         case 'list':
-            type = getPropJsType('object', prop.type.item)
+            type = getPropJsType(imports, 'object', prop.type.item)
             if (type.indexOf('|')) {
                 type = `(${type})[]`
             } else {
@@ -438,6 +457,8 @@ function getScalarJsType(typeName: string): string {
             return 'Date'
         case 'BigInt':
             return 'bigint'
+        case 'BigDecimal':
+            return 'BigDecimal'
         case 'Bytes':
             return 'Uint8Array'
         case 'JSON':
@@ -501,10 +522,10 @@ class ImportRegistry {
     private typeorm = new Set<string>()
     private model = new Set<string>()
     private marshal = false
+    private bigdecimal = false
     private assert = false
 
-    constructor(private owner: string) {
-    }
+    constructor(private owner: string) {}
 
     useTypeorm(...names: string[]): void {
         names.forEach((name) => this.typeorm.add(name))
@@ -521,11 +542,18 @@ class ImportRegistry {
         this.marshal = true
     }
 
+    useBigDecimal() {
+        this.bigdecimal = true
+    }
+
     useAssert() {
         this.assert = true
     }
 
     render(model: Model, out: Output): void {
+        if (this.bigdecimal) {
+            out.line(`import {BigDecimal} from "@subsquid/big-decimal"`)
+        }
         if (this.assert) {
             out.line('import assert from "assert"')
         }