+++ typeorm-codegen/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @subsquid/typeorm-codegen
 
-This log was last generated on Thu, 15 Sep 2022 20:54:52 GMT and should not be manually modified.
+This log was last generated on Thu, 22 Sep 2022 19:06:43 GMT and should not be manually modified.
+
+## 0.2.1
+Thu, 22 Sep 2022 19:06:43 GMT
+
+### Patches
+
+- remove unneeded `big-decimal` dependency
+
+## 0.2.0
+Thu, 22 Sep 2022 15:21:10 GMT
+
+### Minor changes
+
+- support `BigDecimal` type
 
 ## 0.1.2
 Thu, 15 Sep 2022 20:54:52 GMT